Move models to models.client and models.database packages
