+++ app/models/client/UserSummary.scala
@@ -1,10 +1,10 @@
-package models
+package models.client
 
+import ClientJsonFormats._
+import models.database.mongo.{IdentityUser, Orphan, PersistedUser}
 import org.joda.time.DateTime
 import play.api.libs.json.Json
-import repositories.{IdentityUser, Orphan, PersistedUser}
 import scala.language.implicitConversions
-import ClientJsonFormats._
 
 case class UserSummary(id: String,
                        email: String,
