+++ app/models/client/UserUpdateRequestValidator.scala
@@ -1,4 +1,4 @@
-package models
+package models.client
 
 import scalaz.{-\/, \/, \/-}
