+++ app/models/database/postgres/PostgresReservedUsernameRepository.scala
@@ -1,13 +1,13 @@
-package repositories.postgres
+package models.database.postgres
 
 import com.google.inject.Inject
 import com.gu.identity.util.Logging
-import models._
+import models.client
+import models.client.{ApiResponse, ReservedUsername, ReservedUsernameList}
 import play.api.libs.json.Json
 import scalikejdbc._
 
-import scala.concurrent.{ExecutionContext, Future}
-import scalaz.\/
+import scala.concurrent.ExecutionContext
 
 class PostgresReservedUsernameRepository @Inject()(implicit ec: ExecutionContext) extends Logging
   with PostgresJsonFormats
@@ -16,7 +16,7 @@ class PostgresReservedUsernameRepository @Inject()(implicit ec: ExecutionContext
   def loadReservedUsernames: ApiResponse[ReservedUsernameList] = readOnly { implicit session =>
     val sql = sql"SELECT jdoc from reservedusernames"
     val results = sql.map(rs => Json.parse(rs.string(1)).as[ReservedUsername]).list().apply()
-    ReservedUsernameList(results.map(_.username))
+    client.ReservedUsernameList(results.map(_.username))
   }(logFailure("Failed to load reserved usernames"))
 
 }
