+++ test/monitoring/MongoDBHealthCheckTest.scala
@@ -1,13 +1,13 @@
 package monitoring
 
 import akka.actor.ActorSystem
+import models.database.mongo.UsersReadRepository
 import org.scalatest.time.{Millis, Seconds, Span}
 import org.scalatest.{FunSuite, Matchers}
 import org.scalatest.concurrent.Eventually
 import org.scalatest.mockito.MockitoSugar
 import org.mockito.Mockito._
 import org.scalatestplus.play.guice.GuiceOneServerPerSuite
-import repositories.UsersReadRepository
 import play.api.Environment
 
 import scala.concurrent.{ExecutionContext, Future}
